Physical device
