+++ src/Graphics/Graphics.h
@@ -1,3 +1,4 @@
 #pragma once
 
 #include "Instance.h"
+#include "Device.h"
